+++ src/GWindow.hpp
@@ -73,6 +73,7 @@ private:
     void _OnSpawnAction(GSimpleAction *, GVariant *);
     void _OnConnectAction(GSimpleAction *, GVariant *);
     void _OnConnectDlgResponse(Gtk::Dialog &, gint response, Gtk::Builder);
+    void _OnDisconnectAction(GSimpleAction *, GVariant *);
     void _OnSettingsAction(GSimpleAction *, GVariant *);
     void _OnShowMarkupAction(GSimpleAction *, GVariant *);